evaluation/LLMAsAJudge/requirements.txt
anthropic
python-dotenv
openai
google.generativeai
pandas
scipy
matplotlib
seaborn
tabulate